Fixed: replaced np.float with np.float32

+++ anim_utils/motion_editing/fabrik_chain.py
@@ -121,13 +121,13 @@ class FABRIKBone(object):
     def __init__(self, name, child):
         self.name = name
         self.child = child
-        self.position = np.array([0, 0, 0], np.float) # position of joint
+        self.position = np.array([0, 0, 0], np.float32) # position of joint
         self.length = 0
         self.is_root = False
         self.is_leaf = False
 
 
-ROOT_OFFSET = np.array([0,0,0], np.float)
+ROOT_OFFSET = np.array([0,0,0], np.float32)
 
 class FABRIKChain(object):
     def __init__(self, skeleton, bones, node_order, tolerance=0.01, delta_tolerance=0.0001, max_iter=500, frame_offset=3, root_offset=ROOT_OFFSET, activate_constraints=False):
